Update gate jobs
ansible-role-refstack-client [1] has been created in order to make
running refstack-client in automated jobs easier. The role does
basically the same (and more) as the tempestconf-refstack-devstack.yaml
playbook used in zuul jobs to gate changes within refstack-client repo.

After devstack jobs were defined within the role repo by [2] we no
longer need the playbook and the generate-tempestconf-refstack role.
We can reuse the jobs [2] from ansible-role-refstack-client.

The refstack-client-devstack-tempestconf job together with the
playbook and the generate-tempestconf-refstack role can be removed
as python-tempestconf was the only consumer of those, it got removed
from there by the Depends-On patch.

[1] https://opendev.org/x/ansible-role-refstack-client
[2] https://review.opendev.org/c/x/ansible-role-refstack-client/+/765671

Depends-On: https://review.opendev.org/c/osf/python-tempestconf/+/766865
Change-Id: I618c8e59e67f0eed574671d5485bc11f7abec465