Unite tempest_dir in _get_next_stream_subunit_output_file
The method have used tempest_dir value given by the argument as well
as self.tempest_dir value which was confusing and lead to a unit test
failure in the case when tempest_dir was a valid path to an existing
tempest environment.

Usually self.tempest_dir and tempest_dir were the same, the values
were different only in a unittest which was editing the tempest dir
path for test purposes.

The docstring suggests that the method should use the value from the
argument, therefore the patch removes the self.tempest_dir occurrence.

Change-Id: I80dc38e6f0dcff1956f2881894d51a1c35577b33