Inclusive jargon
Following tempest's example where arguments such as --blacklist-file,
--black-regex and --whitelist-file are deprecated by [1], let's do
the change here as well.

A few occurrences cannot be renamed atm as refstack-client uses an older
tempest version which doesn't contain the change [1] yet. After we update
the tempest version used in this project, the rest occurrences will be
renamed as well.

[1] https://review.opendev.org/c/openstack/tempest/+/768583

Change-Id: I514fb688f6895338a877cbc706ddd8d6fc5f906d