Update default tempest tag
The currently used tempest tag points to the version of tempest that
causes failure of some object_storage tests. The failure occurs when
swift is used on ceph. In such a case tempest expects that swift's
API endpoint for listing active capabilities is at /info when
actually the correct endpoint is /swift/info. This issue fixes
commit [1] that should therefore be used as a new default
tempest tag.

[1] https://review.opendev.org/c/openstack/tempest/+/692375

Change-Id: Id213499a54d793fab198a9bc2d96dd4cc3c89ce6