Temporarily disable most of the jobs
refstack jobs are broken and require a fix (see follow up patches)
on refstack-client side which can't be merged without a fix
on ansible-role-refstack-client.

python jobs are broken too and require a fix in refstack-client
(already proposed) which can't be merged without other mentioned
fixes.

This will be reverted as soon as the issue is resolved.

Change-Id: Ic00064ed0769aaad3730f1db6889ca180860370b